% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{calc_metrics}
\alias{calc_metrics}
\title{Calculates metrics for model evaluation.}
\usage{
calc_metrics(X, X_true, eps = 1e-10)
}
\arguments{
\item{X}{DxD matrix of predicted parameters.}

\item{X_true}{DxD matrix of true parameters}

\item{eps}{float. Absolute values below eps will be considered 0.}
}
\description{
Calculates metrics for model evaluation.
}
