% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{censor_network}
\alias{censor_network}
\title{Turns a fully observed network into a partially observed network.}
\usage{
censor_network(G, p)
}
\arguments{
\item{G}{D x D matrix representing direct causal graph.}

\item{p}{Float 0 to 1, proportion of nodes to hide.}
}
\description{
This turns a fully observed network matrix into a partially observed one
where TCEs of missing nodes are integrated into pseudo-DCEs for the observed
nodes.
}
