% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_pca.R
\name{convex_pca}
\alias{convex_pca}
\title{Calculates the convex relaxation of principal component analysis}
\usage{
convex_pca(
  Y,
  r = NULL,
  weights = NULL,
  n_components = NULL,
  its = 1000,
  rmse_target = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{Y}{Matrix. Missing values are allowed.}

\item{r}{Float, sequence of floats, or NULL. Maximum nuclear norm of the
returned solution.}

\item{weights}{Matrix with weights for the loss. If `NULL`, will default
to 1 for every observed value and 0 for every missing value.}

\item{n_components}{Integer or `NULL`. Number of approximate principal
components to keep. If `NULL`, the entire matrix is returned.}

\item{its}{Integer. Maximum number of iterations to run.}

\item{rmse_target}{Float. Stop when rmse change in solution is less than
`rmse_target`.}

\item{verbose}{Boolean. `TRUE` to print convergance progress.}
}
\description{
The convex relaxation of PCA is the optimization problem
  || W(X - Y) ||_F ^ 2 subject to ||X||_* < r
Where W is a weight matrix and ||X||_* is the nuclear (trace) norm of X.
}
\details{
This can be ued to generate a nuclear-norm regularizaed approximation
to Y, even if Y has missing values.
}
