% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_pca.R
\name{convex_pca_worker}
\alias{convex_pca_worker}
\title{Worker function for convex PCA. See `convex_pca` for more.}
\usage{
convex_pca_worker(
  Y,
  r = NULL,
  weights = NULL,
  its = 1000,
  rmse_target = 1e-04,
  warm_start = NULL,
  verbose = FALSE
)
}
\arguments{
\item{Y}{Matrix. See `convex_pca`.}

\item{r}{Float. Maximum nuclear norm of the returned solution.}

\item{weights}{Matrix. See `convex_pca`.}

\item{its}{Integer. See `convex_pca`.}

\item{rmse_target}{Float. See `convex_pca`.}

\item{warm_start}{Matrix. See `convex_pca`}

\item{verbose}{Boolean. See `convex_pca`}
}
\description{
Worker function for convex PCA. See `convex_pca` for more.
}
