% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{cv_fit_regularized}
\alias{cv_fit_regularized}
\title{Fits L1-regularized approximate inverse model.}
\usage{
cv_fit_regularized(R_tce, weights = NULL, k = 1, nfolds = NULL, lambda = NULL)
}
\arguments{
\item{R_tce}{D x D matrix of "total causal effects".}

\item{weights}{Length D vector of per-row weights. Default is 1 for each
observation.}

\item{k}{Number of zero-values to drop during CV.}

\item{nfolds}{Number of folds. NULL for k=1 to use leave-one-out CV.}

\item{lambda}{Vector of lambda values to try.}
}
\description{
Fits L1-regularized approximate inverse model.
}
