% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{delta_cde}
\alias{delta_cde}
\title{Uses delta method to calculate CDE standard error.}
\usage{
delta_cde(R_tce_inv, SE_tce, na.rm = FALSE)
}
\arguments{
\item{R_tce_inv}{DxD matrix. Inverse or approximate inverse of the TCE.}

\item{SE_tce}{DxD matrix. Standard error of the TCE.}
}
\description{
Uses delta method to calculate CDE standard error.
}
