% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{filter_G_on_xi}
\alias{filter_G_on_xi}
\title{Filters estimate of the graph G based on edge inclusion in CV.}
\usage{
filter_G_on_xi(G, xi, thresh = 0.45)
}
\arguments{
\item{G}{D x D matrix}

\item{xi}{D x D matrix}

\item{thresh}{Entries in G with corresponding xi < thresh are set to 0}
}
\description{
Filters estimate of the graph G based on edge inclusion in CV.
}
