% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{filter_tce}
\alias{filter_tce}
\title{Helper function to do basic filtering of the TCE matrix.}
\usage{
filter_tce(R_tce, SE_tce, max_R = 1, max_SE = 0.5, max_nan_perc = 0.5)
}
\arguments{
\item{R_tce}{Matrix or data.frame. Estimates of TCE.}

\item{SE_tce}{Matrix or data.frame. Standard errors of the entries in R_tce.}

\item{max_R}{Float. Set all entries where `abs(R_tce) > max_R` to `NA`.}

\item{max_SE}{Float. Set all entries whwere `SE > max_SE` tp `NA`.}

\item{max_nan_perc}{Float. Remove columns and rows that are more than
`max_nan_perc` NAs.}
}
\description{
Large values of R, entries with a high SE, and row/columns with many nans
can be removed.
}
