% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_pca.R
\name{find_best_r}
\alias{find_best_r}
\title{Cross-validation wrapper for convex PCA worker.}
\usage{
find_best_r(
  Y,
  rs = NULL,
  weights = NULL,
  its = 1000,
  rmse_target = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{Y}{See `convex_pca`.}

\item{rs}{A list of r values to try.}

\item{weights}{See `convex_pca`.}

\item{its}{See `convex_pca`.}

\item{rmse_target}{See `convex_pca`.}

\item{verbose}{Print progress.}
}
\description{
Cross-validation wrapper for convex PCA worker.
}
