% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_direct}
\alias{fit_direct}
\title{Fit simple inspre model}
\usage{
fit_direct(X, Y, lambda = 1, niter = 20, true_R = NULL,
  true_B = NULL)
}
\arguments{
\item{X}{N x M matrix of genotypes.}

\item{Y}{N x D matrix of response.}

\item{lambda}{regularization strength, passed to glmnet.}

\item{niter}{max iterations.}
}
\value{
A list:
  R_hat: D x D matrix of deconvoluted direct effects.
  B_hat: M x D matrix of inferred genotype effect sizes.
  progress_tibble: A tible containing information on convergence.

TODO(brielin): Figure out what to do with the progress of this and why it's
  performance is so poor.
}
\description{
Fit simple inspre model
}
