% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{fit_exact}
\alias{fit_exact}
\title{Fits exact model to data.}
\usage{
fit_exact(R_tce)
}
\arguments{
\item{R_tce}{D x D matrix of "total causal effects".}
}
\value{
D x D matrix with zero diagonal of deconvoluted direct effects.
}
\description{
Fits exact model to data.
}
