% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{fit_glmnet}
\alias{fit_glmnet}
\title{Fit's MVMR model to data.}
\usage{
fit_glmnet(sumstats_fit, snps_to_use)
}
\arguments{
\item{sumstats_fit}{List with elements beta_hat and se_hat.}

\item{snps_to_use}{List of per-phenotype pair SNP to use.}
}
\description{
Fit's MVMR model to data.
}
