% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_ind_level}
\alias{fit_ind_level}
\title{Fits model using individual level genotypes and sumstats for SNP selection.}
\usage{
fit_ind_level(X, Y, sumstats_select, p_thresh = 1e-05, lambda = 0.1,
  niter = 5)
}
\arguments{
\item{X}{N x M matrix of genotypes.}

\item{Y}{N x D matrix of phenotypes.}

\item{sumstats_select}{List with elements "r_squared" and "p_value", each
M x D matrices.}

\item{p_thresh}{p-value threshold for inclusion of a SNP in the fit.}

\item{lambda}{Regularization strength.}

\item{niter}{Number of iterations.}
}
\description{
Fits model using individual level genotypes and sumstats for SNP selection.
}
