% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_inspre}
\alias{fit_inspre}
\title{Fits inverse sparse regression model.}
\usage{
fit_inspre(
  R_tce,
  W = NULL,
  rho = 10,
  lambda = NULL,
  lambda_min_ratio = 0.01,
  nlambda = 20,
  alpha = 0,
  gamma = NULL,
  its = 100,
  delta_target = 1e-04,
  verbose = 1,
  train_prop = 0.8,
  cv_folds = 0,
  mu = 10,
  tau = 2,
  solve_its = 3,
  ncores = 1,
  warm_start = TRUE
)
}
\arguments{
\item{R_tce}{D x D matrix of  "total causal effects".}

\item{W}{DxD Matrix of weights.}

\item{rho}{Float. Initial learning rate for ADMM.}

\item{lambda}{Float, sequence of floats of NULL. L1 regularization strength
on inverse of X. If NULL, a logarithmicallly spaced set of values between
the maximimum absolute off diagonal element of X and lambda_min_ratio
times this value will be used.}

\item{lambda_min_ratio}{Float, ratio of maximum lambda to minimum lambda.}

\item{nlambda}{Integer. Number of lambda values to try.}

\item{alpha}{Float between 0 and 1 or NULL. If > 0, the model will be fit
once with gamma = 0 to find L0, then all subsequent fits will use
gamma = alpha * L0 / D. Set to NULL to provide gamma directly.}

\item{gamma}{Float or sequence of nlambda floats or NULL. Determinant
regularization strength to use (for each lambda value). It is recommended
to set alpha rather than setting this directly.}

\item{its}{Integer. Maximum number of iterations.}

\item{delta_target}{Float. Target change in solution.}

\item{verbose}{0, 1 or 2. 2 to print convergence progress for each lambda,
1 to print convergence result for each lambda, 0 for no output.}

\item{train_prop}{Float between 0 and 1. Proportion of data to use for
training in cross-validation.}

\item{cv_folds}{Integer. Number of cross-validation folds to perform.}

\item{mu}{rho modification parameter for ADMM. Rho will be
increased/decreased when the dual constrant and primal constraint are off
by a factor of > mu.}

\item{tau}{rho modification parameter for ADMM. When called for, rho will be
increased/decreased by the factor tau.}

\item{solve_its}{Integer, number of iterations of bicgstab/lasso to run
for each U and V update.}

\item{ncores}{Integer, number of cores to use.}

\item{warm_start}{Logical. Whether to use previous lambda value result as
starting point for next fit.}
}
\description{
See also inspre::inspre() for more details.
}
