% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_model}
\alias{fit_model}
\title{Fits network mendelian randomization model to data.}
\usage{
fit_model(
  file_pattern_select,
  file_pattern_fit,
  p_thresh = 1e-05,
  shrink = FALSE,
  min_instruments = 5,
  resample = 0
)
}
\arguments{
\item{file_pattern_select}{String representing a file pattern. Sumstats files
matching this pattern will be used for SNP selection.}

\item{file_pattern_fit}{String representing a file pattern. Sumstats files
matching this pattern will be used for SNP selection.}

\item{p_thresh}{Float. p-value threshold for inclusion.}

\item{shrink}{Boolean. True to shrink estimates of R_tce to 0.}

\item{min_instruments}{Integer. Return NA if there are less than
this many instruments for a pair of phenotypes.}

\item{resample}{Non-negative integer or "delta". The number of resample iterations.}
}
\description{
Fits network mendelian randomization model to data.
}
