% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{fit_mrbma}
\alias{fit_mrbma}
\title{Fit's MVMR model to data.}
\usage{
fit_mrbma(sumstats_fit, snps_to_use, max_iter = 1000, verbose = FALSE)
}
\arguments{
\item{sumstats_fit}{List with elements beta_hat and se_hat.}

\item{snps_to_use}{List of per-phenotype pair SNP to use.}

\item{max_iter}{Integer. Number of BMA iterations to run, default 1000.}

\item{verbose}{Boolean. True to print progress.}
}
\description{
Fit's MVMR model to data.
}
