% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_regularized}
\alias{fit_regularized}
\title{Fits L1-regularized approximate inverse model.}
\usage{
fit_regularized(R_tce, lambda = 0.1)
}
\arguments{
\item{R_tce}{D x D matrix of "total causal effects".}

\item{lambda}{Float. Regularization strength.}
}
\description{
Fits L1-regularized approximate inverse model.
}
