% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{fit_regularized_cheat}
\alias{fit_regularized_cheat}
\title{Fits L1-regularized approximate inverse while finding best lambda (cheating).}
\usage{
fit_regularized_cheat(R_tce, R)
}
\arguments{
\item{R_tce}{DxD matrix of "total causal effects".}

\item{R}{DxD matrix of true "causal direct effects".}
}
\description{
Fits L1-regularized approximate inverse while finding best lambda (cheating).
}
