% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_regularized_cv}
\alias{fit_regularized_cv}
\title{Fits L1-regularized approximate inverse model.}
\usage{
fit_regularized_cv(
  R_tce,
  weights = NULL,
  k = 1,
  nfolds = NULL,
  lambda = c(0.1, 0.05, 0.01, 0.005, 0.001, 5e-04, 1e-04)
)
}
\arguments{
\item{R_tce}{D x D matrix of "total causal effects".}

\item{weights}{Length D vector of per-row weights. Default is 1 for each observation.}

\item{k}{Number of zero-values to drop during CV.}

\item{nfolds}{Number of folds. NULL for k=1 to use leave-one-out CV.}

\item{lambda}{Vector of lambda values to try.}
}
\description{
Fits L1-regularized approximate inverse model.
}
