% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_tce}
\alias{fit_tce}
\title{Selects SNPs for inclusion in MR by comparing per-variance effect sizes.}
\usage{
fit_tce(
  sumstats,
  selected_snps,
  mr_method = "egger_w",
  min_instruments = 5,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sumstats}{List representing summary statistics from the second
dataset. Must include entries "beta_hat" and "se_hat".}

\item{selected_snps}{A list of lists with names of each equal to the
phenotype names. The inner lists are boolean vectors of length equal to
the number of SNPs and TRUE indicating to use that SNP.}

\item{mr_method}{String, one of c("mean", "raps"). Method to use for TCE
estimate between every pair.}

\item{min_instruments}{Integer. Return NA if there are less than
this many instruments for a pair of phenotypes.}

\item{verbose}{Bpplean. True to print progress.}

\item{...}{Additional parameters to pass to mr_method.}

\item{snps_to_use}{List or NULL. A list named by phenotypes where each list
entry is a list of SNPs that can be used for that phenotype. Usually
the result of clumping to avoid correlated SNPs.}

\item{p_thresh}{Float, p-value threshold to use for SNP inclusion.}

\item{exclusive}{Bool. True to only use SNPs significant for one phenotype
but *not* the other.}

\item{weight}{Bool. True to store welch-test weights for regression.}

\item{filter}{Double of NULL. If not NULL, filter variants with welch
statistic less than filter.}
}
\description{
Notes: sumstats passed to this function must be computed on the per-variance
scale.
}
