% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_beta}
\alias{generate_beta}
\title{Generates sparse effects and Mendelian randomization conditioning statistic.}
\usage{
generate_beta(M, D, p = 0.1, pleiotropy = FALSE)
}
\arguments{
\item{M}{Integer. Number of SNPs to simulate.}

\item{D}{Integer. Number of phenotypes to simulate.}

\item{p}{Float. Proportion of non-zero effects.}

\item{pleiotropy}{Bool. TRUE to allow for SNPs to effect multiple phenotypes.}
}
\value{
A list,
  beta: sparse float matrix of dimension MxD.
  condition: float vector of length D.
}
\description{
Generates an MxD matrix of SNP effects with p% non-zero in expectation. Also
returns a vector of MR conditioning, the maximum ratio of the effect size of
any SNP on each gene relative to every other gene.
}
