% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_beta_pair}
\alias{generate_beta_pair}
\title{Generates pairs of optionally correleated SNP effects.}
\usage{
generate_beta_pair(M_s, M_p, rho)
}
\arguments{
\item{M_s}{Integer. Number of shared SNPs.}

\item{M_p}{Integer or sequence of two integers. Number of private SNPs to
simulate per phenotype pair.}

\item{rho}{Float [-1, 1]. Correlation of shared SNPs.}
}
\description{
Generates pairs of optionally correleated SNP effects.
}
