% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_confounding}
\alias{generate_confounding}
\title{Generates (optionally correlated) environmental confounding.}
\usage{
generate_confounding(N, C, D, sigma_g)
}
\arguments{
\item{N}{Integer, number of individuals.}

\item{C}{Integer, dimensionality of confounding.}

\item{D}{Integer, number of phenotypes.}

\item{sigma_g}{Float or D x D matrix of floats. If a single value it will be
taken as the variance of the generated effect size. If a CxC matrix it will
be the covariance.}
}
\description{
Note that this is different from the (always uncorrelated) measurement error
since it is mixed by the network with the genetic component.
}
