% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generates dataset.}
\usage{
generate_dataset(N, D, G, M, p, h, censor_prob = 0, beta = NULL)
}
\arguments{
\item{N}{Integer or array of integers of length D. Number of individuals to simulate.}

\item{D}{Integer. Number of phenotypes to simulate.}

\item{G}{D x D matrix. Causal graph to simulate.}

\item{M}{Integer. Number of SNPs to simulate.}

\item{p}{Float between 0 and 1. Per-SNP effect probability.}

\item{h}{Float between 0 and 1. Phenotype heritability.}

\item{censor_prob}{Float between 0 and 1. Hide phenotypes to induce
non-causal genetic correlation. Proportion of phenotype to hide.}

\item{beta}{M x D matrix of SNP effects or NULL to generate the effects based
on the provided parameters.}
}
\description{
Generates a network mendelian randomization dataset with observed phenotypes,
genotypes, true network effects and true effect sizes.
}
