% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_genotypes}
\alias{generate_genotypes}
\title{Generates genotypes.}
\usage{
generate_genotypes(N, M)
}
\arguments{
\item{N}{Integer. Number of individuals to simulate.}

\item{M}{Integer. Number of SNPs to simulate.}
}
\value{
and NxM matrix of simulated genotypes.
}
\description{
Generates genotypes.
}
