% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_network}
\alias{generate_network}
\title{Generates and optionally normalizes causal graph.}
\usage{
generate_network(
  D,
  graph = "random",
  orient = "random",
  v = 0.3,
  v_max = 0.9,
  v_min = 0.01,
  pert = F
)
}
\arguments{
\item{D}{Integer. Number of phenotypes to simulate.}

\item{graph}{String, one of "random", "hub" or "scale-free". Specifies the
kind of graph to be generated.}

\item{orient}{String, one of "random", "towards" or "away". Specifies the
edge orientation strategy. Randomly, towards high degree nodes, or away
from high degree nodes.}

\item{v}{Float between 0 and 1. Roughly corresponds to edge weight, see
huge.generator for more information. Default 0.3.}

\item{v_max}{Float greater than v. Used as max value if pert=T. Default 0.9.}

\item{v_min}{Float greater than 0. Minimim value of pert=T. Default 0.01.}

\item{pert}{Boolean. TRUE to replace huge produced value with one sampled
from the betaPERT distribution with mode v and max v_max.}
}
\description{
This is a wrapper around the huge.generator function from that package. After
generating the adjacency matrix, this orients the sign randomly and then the
edges according to the specified method.
}
