% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_sumstats}
\alias{generate_sumstats}
\title{Generate summary statistics.}
\usage{
generate_sumstats(X, Y)
}
\arguments{
\item{X}{N x M matrix of genotypes.}

\item{Y}{N x D matrix of phenotypes.}
}
\value{
A list,
  beta_hat: An M x D matrix of calculated effect sizes.
  se_hat: An M x D matrix of corresponding estimated standard errors.
  p_value: An M x D matrix of corresponding p-values.
  r_squared: An M x D matrix of variances explained.
}
\description{
Generate summary statistics.
}
