% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_direct}
\alias{get_direct}
\title{Gets direct network from observed effects.}
\usage{
get_direct(R_obs)
}
\arguments{
\item{R_obs}{D x D matrix of direct effects.}
}
\value{
D x D matrix of observed effects.
}
\description{
Gets direct network from observed effects.
}
