% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{get_observed}
\alias{get_observed}
\title{Gets observed network from direct effects.}
\usage{
get_observed(R)
}
\arguments{
\item{R}{D x D matrix of direct effects.}
}
\value{
D x D matrix of observed effects.
}
\description{
Gets observed network from direct effects.
}
