% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{get_tce}
\alias{get_tce}
\title{Gets matrix of TCE from observed network.}
\usage{
get_tce(R_obs)
}
\arguments{
\item{R_obs}{D x D matrix of observed effects.}
}
\value{
D x D matrix of total causal effects.
}
\description{
Gets matrix of TCE from observed network.
}
