% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_pca.R
\name{make_weights}
\alias{make_weights}
\title{Calculates weights to use for convex PCA from observed SEs.}
\usage{
make_weights(SE, max_weight = NULL)
}
\arguments{
\item{SE}{A matrix of floats, with missing entries (NA) allowed.}

\item{max_weight}{The maximum allowable weight to use. Entries with 1/SE^2
above `max_weight` will get weight `max_weight`.}
}
\description{
This uses standard inverse variance weighting for 0 for missing entries.
If there are entries with SE 0, they will receive the same weight as the
largest weighted non-zero entry, rather than Infinity.
}
