% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{naive_ma}
\alias{naive_ma}
\title{Uses naive meta analysis method for TCE estimation with multiple SNPs.}
\usage{
naive_ma(beta_exp, beta_out, se_exp, se_out)
}
\arguments{
\item{beta_exp}{Vector of SNP effects on exposure variable.}

\item{beta_out}{Vector of SNP effects on outcome variable.}

\item{se_exp}{Vector of SNP effects on exposure variable.}

\item{se_out}{Vector of SNP effects on outcome variable.}
}
\value{
A list with two elements.
  tce_hat: The total causal efffect estimate.
  se_tce: The standard error of the estimate.
}
\description{
Uses naive meta analysis method for TCE estimation with multiple SNPs.
}
