% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{parse_sumstats_multi_trait}
\alias{parse_sumstats_multi_trait}
\title{Parses a list of sumstats tibbles, returning a list of matrices.}
\usage{
parse_sumstats_multi_trait(sumstats_list, min_maf = 0.001)
}
\arguments{
\item{sumstats_list}{A list where each entry is a tibble of sumstats.}

\item{min_maf}{Float, minimum allele frequency to keep.}
}
\description{
This parses a list of sumstats tibbles into a list of matrices containing
the relevant data for future analysis, ie that required by `fit_sumstats`.
}
