% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{parse_sumstats_neale}
\alias{parse_sumstats_neale}
\title{Reads single sumstats file in the format of the Neale lab UKBB analysis.}
\usage{
parse_sumstats_neale(sumstats)
}
\arguments{
\item{sumstats}{Tibble. Must include ccolumns "minor_AF",
"low_confidence_variant", "tstat", "n_complete_samples", "beta", "se"
and "pval".}
}
\description{
Note that this converts beta values to a normalized scale if they are not
already on one. Note also that this reads the entire dataset into memory.
If the number of SNPs is
large, consider doing some preprocessing of your data to remove SNPs that
are unlikely to be used in the analysis.
Note also that this only keeps the normalized scale beta and SE due to memory
concerns. The number of samples and p-value can be backed out from these.
}
