% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_files}
\alias{read_files}
\title{Wrapper for read_gz_tsv to read all files matching pattern.}
\usage{
read_files(file_pattern = NULL, file_list = NULL)
}
\arguments{
\item{file_pattern}{A string representing a glob pattern of files to match.}

\item{file_list}{A list of files to open.}
}
\description{
One of file_pattern and file_list must be provided.
}
