% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_gz_tsv}
\alias{read_gz_tsv}
\title{Simple wrapper for readr::read_tsv to use a gzipped file.}
\usage{
read_gz_tsv(filename)
}
\arguments{
\item{filename}{A string ending in .tsv.gz or .tsv.bgz to be read.}
}
\description{
Simple wrapper for readr::read_tsv to use a gzipped file.
}
