% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_snp_list}
\alias{read_snp_list}
\title{Reads a file pattern corresponding to files with potential instruments.}
\usage{
read_snp_list(file_pattern)
}
\arguments{
\item{file_pattern}{A file pattern corresponding to a set of files, one for
each phenotype. Each file should contain a list of potential SNPs to use as
instruments for that phenotype.}
}
\description{
Reads a file pattern corresponding to files with potential instruments.
}
