% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_sumstats_neale}
\alias{read_sumstats_neale}
\title{Reads a pattern of summary statistics formatted according to the Neale lab.}
\usage{
read_sumstats_neale(
  file_pattern = NULL,
  file_list = NULL,
  chunk_size = 20,
  verbose = TRUE
)
}
\arguments{
\item{file_pattern}{A glob string specifying the files to read or NULL to
use file_list instread.}

\item{file_list}{A list of file names to use, or NULL to use file_pattern
instead.}

\item{chunk_size}{Number of phenotypes to pivot at a time.}

\item{verbose}{Boolean. True to print progress.}
}
\description{
Reads a pattern of summary statistics formatted according to the Neale lab.
}
