% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_ukbbss_neale}
\alias{read_ukbbss_neale}
\title{Reads a pattern of summary statistics formatted according to the Neale lab.}
\usage{
read_ukbbss_neale(file_pattern, min_maf = 0.001)
}
\arguments{
\item{file_pattern}{A glob string specifying the files to read.}

\item{min_maf}{Float, minimum allele frequency to keep.}
}
\description{
Reads a pattern of summary statistics formatted according to the Neale lab.
}
