% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{resample_cde}
\alias{resample_cde}
\title{Resample CDE estimate using observed standard errors of TCE.}
\usage{
resample_cde(R_tce, SE_tce, fit_method_func, niter = 100)
}
\arguments{
\item{R_tce}{DxD matrix of floats. Estimates of R_tce to resample.}

\item{SE_tce}{DxD matrix of floats. Standard errors of entries in R_tce.}

\item{fit_method_func}{Function. Method for inferring R_cde given R_tce.
Must return a list with entry "R_hat".}

\item{niter}{Integer. Number of resampling iterations.}
}
\description{
Resample CDE estimate using observed standard errors of TCE.
}
