% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{resample_tce}
\alias{resample_tce}
\title{Resample TCE estimates using observed SEs while imputing missing values.}
\usage{
resample_tce(
  R_tce,
  SE_tce,
  niter = 100,
  impute_function = NULL,
  rmse_target = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{R_tce}{Matrix.}

\item{SE_tce}{Matrix, same dimensnions as `R_tce`.}

\item{niter}{Integer, number of resampling iterations.}

\item{impute_function}{Function to use to impute values of R_tce.}

\item{rmse_target}{Float. Target RMSE change per iteration to stop.}

\item{verbose}{Bool. True to print progress.}
}
\description{
Resample TCE estimates using observed SEs while imputing missing values.
}
