% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{welch_filter_both_sig}
\alias{welch_filter_both_sig}
\title{Wrapper for `welch_test` that sets SNPs significant for both to FALSE.}
\usage{
welch_filter_both_sig(beta1, se1, beta2, se2, sig2, welch_thresh = 0.05)
}
\arguments{
\item{beta1}{Float or vector of floats, mean of the first sample.}

\item{se1}{Float or vector of floats, SD of estimate of mu1.}

\item{beta2}{Float or vector of floats, mean of the second sample.}

\item{se2}{Float or vector of floats, SD of estimate of mu2.}

\item{sig2}{Bool or vector of bools indicating whether this SNP is also
significant for phenotype two.}

\item{welch_thresh}{Float, p_value threshold for significance.}
}
\description{
Wrapper for `welch_test` that sets SNPs significant for both to FALSE.
}
