% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{count_instruments}
\alias{count_instruments}
\title{A simple helper function to count the number of instuments for each pair.}
\usage{
count_instruments(selected)
}
\arguments{
\item{selected}{A list of lists, output of `select_snps`.}
}
\description{
A simple helper function to count the number of instuments for each pair.
}
