% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{filter_tce}
\alias{filter_tce}
\title{Helper function to do basic filtering of the TCE matrix.}
\usage{
filter_tce(
  R_tce,
  SE_tce,
  R_tce_true = NULL,
  max_R = 1.5,
  max_SE = 3,
  max_nan_perc = 0.5
)
}
\arguments{
\item{R_tce}{Matrix or data.frame. Estimates of TCE.}

\item{SE_tce}{Matrix or data.frame. Standard errors of the entries in R_tce.}

\item{R_tce_true}{Matrix or data.frame. For simulation and testing, remove
columns/rows from the true R_tce as well as our estimate.}

\item{max_R}{Float. Set all entries where `abs(R_tce) > max_R` to `NA`.}

\item{max_SE}{Float. Set all entries where `SE_tce > max_SE` to `NA`.}

\item{max_nan_perc}{Float. Remove columns and rows that are more than
`max_nan_perc` NAs.}
}
\description{
Large values of R, entries with a high SE, and row/columns with many nans
can be removed.
}
