% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_model}
\alias{fit_model}
\title{Fits network mendelian randomization model to data.}
\usage{
fit_model(sumstats_select, sumstats_fit, mr_method = c("mean", "raps"),
  fit_method = c("exact", "regularized"), p_thresh = 1e-05)
}
\arguments{
\item{sumstats_select}{List representing summary statistics from the first
dataset. Must include "p-value" and "r-squared".}

\item{sumstats_fit}{List representing summary statistics from the second
dataset. Must include entries "beta_hat" and "se_hat".}

\item{mr_method}{String, one of c("mean", "raps"). Method to use for TCE
estimate between every pair.}

\item{fit_method}{String, one of c("exact", "regularized"). Method to use for
network optimization.}

\item{p_thresh}{Float. p-value threshold for inclusion.}
}
\description{
Fits network mendelian randomization model to data.
}
