% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{fit_tce}
\alias{fit_tce}
\title{Calculates matrix of total causal effects using a specified method.}
\usage{
fit_tce(sumstats_select, sumstats_fit, mr_method = c("mean", "raps"),
  p_thresh = 1e-05, ...)
}
\arguments{
\item{sumstats_select}{List representing summary statistics from the first
dataset. Must include "p-value" and "r-squared".}

\item{sumstats_fit}{List representing summary statistics from the second
dataset. Must include entries "beta_hat" and "se_hat".}

\item{mr_method}{String, one of c("mean", "raps"). Method to use for TCE
estimate between every pair.}

\item{p_thresh}{Float. p-value threshold for inclusion.}

\item{...}{Additional parameters to pass to mr_method.}
}
\description{
Calculates matrix of total causal effects using a specified method.
}
