% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generates dataset.}
\usage{
generate_dataset(N, M, D, p_beta = 0.2, p_net = 0.2, noise = 0,
  pleiotropy = FALSE)
}
\arguments{
\item{N}{Integer. Number of individuals to simulate.}

\item{M}{Integer. Number of SNPs to simulate.}

\item{D}{Integer. Number of phenotypes to simulate.}

\item{p_beta}{Float. Proportion of SNPs with non-zero effect size.}

\item{p_net}{Float. Proportion of non-zero network edges.}

\item{noise}{Float. Proprotion of variance in phenotype attributable to
noise. 0 for no noise, 1 for all noise with no genetic/network component.}

\item{pleiotropy}{Bool. TRUE to allow for SNPs to effect multiple phenotypes.}
}
\description{
Generates a network mendelian randomization dataset with observed phenotypes,
genotypes, true network effects and true effect sizes.
}
