% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_genotypes}
\alias{generate_genotypes}
\title{Generates genotypes.}
\usage{
generate_genotypes(N, M, whiten = FALSE)
}
\arguments{
\item{N}{Integer. Number of individuals to simulate.}

\item{M}{Integer. Number of SNPs to simulate.}

\item{whiten}{Bool. TRUE to whiten such that the observed covariance is
exactly 0. This should only be used for testing.}
}
\value{
and NxM matrix of simulated genotypes.
}
\description{
Generates genotypes.
}
