% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_network}
\alias{generate_network}
\title{Generates and optionally normalizes causal graph.}
\usage{
generate_network(
  D,
  p = 0.1,
  normalize = TRUE,
  epsilon = 0.1,
  symmetric = NULL,
  sd = 1
)
}
\arguments{
\item{D}{Integer. Number of phenotypes to simulate.}

\item{p}{Float. Proportion of non-zero edges.}

\item{normalize}{Bool. TRUE to normalize the generated matrix to have
eigenvalues with modulus between -1 and 1.}

\item{epsilon}{Float. Number to add to maximum eigenvalue to better-condition
normalization.}

\item{symmetric}{Bool. TRUE to force the returned matrix to be
symmetric, false for anti-symmetric, and NULL for no symmetry enforcement.}

\item{sd}{Float. Standard deviation of network edge weights. NA for binary
matrices with equal probability of}
}
\value{
A DxD sparse matrix.
}
\description{
Generates a DxD sparse matrix with normally distributed edge weights.
Optionally normalize the matrix to have eigenvalues with modulus between
-1 and 1.
}
