% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_sumstats}
\alias{generate_sumstats}
\title{Generate summary statistics.}
\usage{
generate_sumstats(X, Y)
}
\arguments{
\item{X}{N x M matrix of genotypes.}

\item{Y}{N x D matrix of phenotypes.}
}
\value{
A list,
  beta_hat: An M x D matrix of calculated effect sizes.
  se_hat: An M x D matrix of corresponding estimated standard errors.
}
\description{
Note that this always returns values on the normalized (per-variance) scale.
}
