% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_sumstats}
\alias{generate_sumstats}
\title{Generates sumstats.}
\usage{
generate_sumstats(beta, M_null, N)
}
\arguments{
\item{beta}{M x D matrix of effect sizes.}

\item{M_null}{Integer. Adds M_null additional SNPs with null effects.}

\item{N}{Integer or array of integers of length D. Number of samples for each
phenotype.}
}
\description{
Generates sumstats.
}
