% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_file_pattern}
\alias{read_file_pattern}
\title{Wrapper for read_gz_tsv to read all files matching pattern.}
\usage{
read_file_pattern(file_pattern)
}
\arguments{
\item{file_pattern}{A string representing a glob pattern of files to match.}
}
\description{
Wrapper for read_gz_tsv to read all files matching pattern.
}
