% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{read_ukbbss_neale}
\alias{read_ukbbss_neale}
\title{Reads a pattern of summary statistics formatted according to the Neale lab.}
\usage{
read_ukbbss_neale(file_pattern)
}
\arguments{
\item{file_pattern}{A glob string specifying the files to read.}
}
\description{
Reads a pattern of summary statistics formatted according to the Neale lab.
}
