% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{select_snps}
\alias{select_snps}
\title{Selects SNPs for inclusion in MR by comparing per-variance effect sizes.}
\usage{
select_snps(
  sumstats,
  snps_to_use = NULL,
  p_thresh = 5e-06,
  exclusive = TRUE,
  weight = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sumstats}{List with elements "beta_hat", "se_hat", both M x D matrices.}

\item{snps_to_use}{List or NULL. A list named by phenotypes where each list
entry is a list of SNPs that can be used for that phenotype. Usually
the result of clumping to avoid correlated SNPs.}

\item{p_thresh}{Float, p-value threshold to use for SNP inclusion.}

\item{exclusive}{Bool. True to only use SNPs significant for one phenotype
but *not* the other.}

\item{weight}{Bool. True to store welch-test weights for regression.}

\item{verbose}{Bool. If true, print phenotype label during iteration.}
}
\description{
Notes: sumstats passed to this function must be computed on the per-variance
scale.
}
