% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{select_snps}
\alias{select_snps}
\title{Selects SNP sets for each phenotype.}
\usage{
select_snps(sumstats_select, p_thresh = 1e-05)
}
\arguments{
\item{sumstats_select}{List with elements "r_squared" and "p_value", each
M x D matrices.}

\item{p_thresh}{Float, p-value threshold to use for SNP inclusion.}
}
\value{
A list with names from the columns of r_squared (phenotypes), where
  each entry is a vector with names from the rows (SNPs) consististing of the
  r_squared value for each SNP that has a significant effect on that
  phenotype accoring to p_thresh.
}
\description{
Selects SNP sets for each phenotype.
}
