% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{select_snps_oracle}
\alias{select_snps_oracle}
\title{Selects SNPs based on true effect sizes.}
\usage{
select_snps_oracle(beta_true)
}
\arguments{
\item{beta_true}{M x D matrix of true effect sizes.}
}
\description{
Selects SNPs based on true effect sizes.
}
