% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{shrink_R}
\alias{shrink_R}
\title{Shrinks the estimate of the total causal effect matrix.}
\usage{
shrink_R(R_tce, SE_tce, lambda = NULL)
}
\arguments{
\item{R_tce}{A matrix of floats with diagonals equal to 1.0. The estimate
of R_tce to be shunk.}

\item{SE_tce}{A matrix of floats, standard errors of the entries in `R_tce`.}

\item{lambda}{Float less than 1.0 or NULL. Amount of shrinkage. NULL
computes shrinkage automatically.}
}
\description{
Shrinks the estimate of the total causal effect matrix.
}
