% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{welch_test}
\alias{welch_test}
\title{Simple implementation of a welch test.}
\usage{
welch_test(mu1, mu2, s1, s2, n1, n2)
}
\arguments{
\item{mu1}{Float, mean of the first sample.}

\item{mu2}{Float, mean of the second sample.}

\item{s1}{Float, SD of estimate of mu1.}

\item{s2}{Float, SD of estimate of mu2.}

\item{n1}{Integer, number of samples in dataset 1.}

\item{n2}{Integer, number of samples in dataset 2.}
}
\description{
This tests the null hypothesis mu1 = mu2 against the one-sided alternative
mu1 > mu2.
}
